Create Question 1
Write a trigger to count all contacts related to the same record type and update the Account as per the record type of contact.
Example -> Contact Record Type - > Student, Teacher
           Contact 1 -> Student
           Contact 2 -> Teacher
           Contact 3 -> Student
           Contact 4 -> Student
There are two fields on Account Total number of Students and Total number of Teachers. Need to update both fields as per the record of contact related to the Account.